Finalise the working component
